RemoveUser

+++ src/main/java/AccountManager.java
@@ -69,14 +69,43 @@ public class AccountManager {
         st.executeUpdate();
     }
 
-    public boolean removeAcc(String name) throws SQLException {
+    public int removeAcc(String name) throws SQLException {
         if(hasAcc(name)){
-            String query = "DELETE FROM users WHERE username = ?";
-            PreparedStatement st = con.prepareStatement(query);
-            st.setString(1, name);
-            st.executeUpdate();
-            return true;
+            if(!isAdmin(name)) {
+                String query = "DELETE FROM users WHERE username = ?";
+                PreparedStatement st = con.prepareStatement(query);
+                st.setString(1, name);
+                st.executeUpdate();
+                return 1;
+            }else{
+                return 2;
+            }
+        }
+        return 0;
+    }
+    public boolean isAdmin(String name) throws SQLException {
+        String query = "SELECT * FROM users WHERE username = ?";
+        PreparedStatement st = con.prepareStatement(query);
+        st.setString(1 , name);
+        ResultSet rs = st.executeQuery();
+        if(rs.next()){
+            return rs.getInt(3) == 1;
         }
         return false;
+    }
+
+    public int promoteToAdmin(String name) throws SQLException {
+        if(hasAcc(name)){
+            if(isAdmin(name)){
+                return 2;
+            }else{
+                String query = "UPDATE users SET is_admin = 1 WHERE username = ?";
+                PreparedStatement st = con.prepareStatement(query);
+                st.setString(1, name);
+                st.executeUpdate();
+                return 1;
+            }
+        }
+        return 0;
     }
 }
